Continually changing data class structure for a better json structure

+++ src/main/java/org/projectfk/blog/data/PostRepo.kt
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface PostRepo : CrudRepository<Blog, Long> {
+interface PostRepo : CrudRepository<Blog, Int> {
 
     fun findByAuthor(user: User) : Iterable<Blog>
